embedding_engine.py
from sentence_transformers import SentenceTransformer
import faiss
import numpy as np

model = SentenceTransformer("all-MiniLM-L6-v2")

def embed_text(text):

    vector = model.encode([text])
    return vector