requirements.txt
streamlit
openai
langchain
faiss-cpu
pypdf
python-dotenv
sentence-transformers